# openapi.yaml
openapi: 3.1.0
info:
  title: BearStack HTTP API
  version: 0.24.0
  description: |
    Programmatic HTTP interface for documents, uploads, tags, custom fields,
    virtual folders and document media.

    BearStack accepts HTTP Basic authentication with the configured credentials.
    A browser session cookie obtained through the web login is accepted as well.
    When authentication is disabled in the server configuration, the security
    requirements are not enforced.

    Write operations require the capabilities stated in their descriptions.
    API clients using form endpoints should send `Accept: application/json`.
servers:
  - url: /
    description: Current BearStack instance
security:
  - basicAuth: []
  - sessionCookie: []
...
paths:
...
  /photos/people:
    get:
      tags:
      - Photos
      operationId: listPhotoPeople
      summary: List automatically grouped people
      description: 'Requires photos.read. HTML by default; format=json or Accept: application/json
        returns a page of at most 60 groups. Admin-only photos are excluded.'
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: q
        in: query
        schema:
          type: string
      - name: format
        in: query
        schema:
          type: string
          enum:
          - json
        description: Request JSON instead of HTML.
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoPeoplePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /settings/photos/faces:
    get:
      tags:
      - Photos
      operationId: getFaceRecognitionSettings
      summary: Recognition settings and progress
      description: Requires photos.manage. HTML by default; format=json returns settings
        and status. Service tokens and vectors are never returned.
      parameters:
      - name: format
        in: query
        schema:
          type: string
          enum:
          - json
        description: Request JSON instead of HTML.
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaceSettingsStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
      - Photos
      operationId: saveFaceRecognitionSettings
      summary: Configure optional face recognition
      description: Requires photos.manage. Form writes redirect to the corresponding
        HTML page. Existing authentication, same-origin protection and audit logging
        apply.
      parameters: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                enabled:
                  type: string
                  enum:
                  - '1'
                  description: Omit to pause. Activation requires a healthy, compatible
                    configured service.
                batch_size:
                  type: integer
                  minimum: 1
                  maximum: 1000
                  default: 100
                delay_millis:
                  type: integer
                  minimum: 100
                  maximum: 60000
                  default: 1000
                interval_minutes:
                  type: integer
                  minimum: 1
                  maximum: 1440
                  default: 15
              required: []
      responses:
        '303':
          description: Saved; Location points to the corresponding page.
          headers:
            Location:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Credentials configured for the BearStack instance.
    sessionCookie:
      type: apiKey
      in: cookie
      name: bearstack_session
      description: Signed session cookie created by the web login.
...
  responses:
    BadRequest:
      description: Invalid request parameters or form data.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid authentication.
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Basic realm="BearStack", charset="UTF-8"
      content:
        text/plain:
          schema:
            type: string
    Forbidden:
      description: The authenticated account lacks the required capability.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Requested resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RecognizedPhotoFace:
      type: object
      required:
      - id
      - person_id
      - name
      - path
      - x
      - y
      - width
      - height
      - manual
      - ignored
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        person_id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
        path:
          type: string
        x:
          type: number
          minimum: 0
          maximum: 1
        y:
          type: number
          minimum: 0
          maximum: 1
        width:
          type: number
          minimum: 0
          maximum: 1
        height:
          type: number
          minimum: 0
          maximum: 1
        manual:
          type: boolean
        ignored:
          type: boolean
    PhotoPerson:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
        count:
          type: integer
          description: Distinct visible photos.
        face_id:
          type: integer
          format: int64
          minimum: 1
    PhotoPeoplePage:
      type: object
      properties:
        query:
          type: string
        people:
          type: array
          items:
            $ref: '#/components/schemas/PhotoPerson'
        faces:
          type: array
          items:
            $ref: '#/components/schemas/RecognizedPhotoFace'
        person_id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
        page:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
...
    FaceSettingsStatus:
      type: object
      properties:
        settings:
          type: object
          properties:
            enabled:
              type: boolean
            batch_size:
              type: integer
            delay_millis:
              type: integer
            interval_minutes:
              type: integer
        status:
          type: object
          properties:
            errors:
              type: array
              maxItems: 20
              items:
                type: object
                properties:
                  path:
                    type: string
                  error:
                    type: string
                  attempts:
                    type: integer
            queued:
              type: integer
              minimum: 0
            done:
              type: integer
              minimum: 0
            failed:
              type: integer
              minimum: 0
            faces:
              type: integer
              minimum: 0
            people:
              type: integer
              minimum: 0
        running:
          type: boolean
        configured:
          type: boolean
        error:
          type: string